fix: some more improvement for core folder
