+++ apps/gauzy/src/app/@core/services/index.ts
@@ -1,3 +1,9 @@
+export * from './layout.service';
+export * from './seo.service';
+export * from './ripple.service';
+export * from './analytics/analytics.service';
+export * from './analytics/jitsu.service';
+
 export * from './accounting-template.service';
 export * from './appointment-employees.service';
 export * from './approval-policy.service';
