+++ apps/gauzy/src/app/@core/services/selector.service.ts
