+++ apps/gauzy/src/app/@core/utils/index.ts
@@ -1,15 +1,3 @@
-import { LayoutService } from './layout.service';
-import { AnalyticsService } from './analytics.service';
-import { PlayerService } from './player.service';
-import { StateService } from './state.service';
-import { SeoService } from './seo.service';
 import { ColorAdapter } from './color-adapter';
 
-export {
-	LayoutService,
-	AnalyticsService,
-	PlayerService,
-	SeoService,
-	StateService,
-	ColorAdapter,
-};
+export { ColorAdapter };
