+++ apps/gauzy/src/app/@shared/sidebar-menu/sidebar-menu.service.ts
@@ -1,17 +1,17 @@
 import { Injectable } from '@angular/core';
 
 @Injectable({
-  providedIn: 'root'
+	providedIn: 'root'
 })
 export class SidebarMenuService {
-  private _selectedItem: any;
-  constructor() { }
+	private _selectedItem: any;
+	constructor() {}
 
-  public get selectedItem(): any {
-    return this._selectedItem;
-  }
+	public get selectedItem(): any {
+		return this._selectedItem;
+	}
 
-  public set selectedItem(value: any) {
-    this._selectedItem = value;
-  }
+	public set selectedItem(value: any) {
+		this._selectedItem = value;
+	}
 }
