+++ apps/gauzy/src/app/@theme/components/header/header.component.ts
@@ -1,18 +1,6 @@
-import {
-	Component,
-	Input,
-	OnDestroy,
-	OnInit,
-	AfterViewInit
-} from '@angular/core';
+import { Component, Input, OnDestroy, OnInit, AfterViewInit, ChangeDetectorRef } from '@angular/core';
 import { Router, NavigationEnd } from '@angular/router';
-import {
-	NbSidebarService,
-	NbThemeService,
-	NbMenuItem,
-	NbDialogService,
-	NbDialogRef
-} from '@nebular/theme';
+import { NbSidebarService, NbThemeService, NbMenuItem, NbDialogService, NbDialogRef } from '@nebular/theme';
 import { combineLatest, firstValueFrom, lastValueFrom, Subject } from 'rxjs';
 import { debounceTime, filter, tap } from 'rxjs/operators';
 import { TranslateService } from '@ngx-translate/core';
@@ -48,9 +36,8 @@ import {
 	UsersOrganizationsService
 } from '../../../@core/services';
 import { DEFAULT_SELECTOR_VISIBILITY, ISelectorVisibility, SelectorBuilderService } from '../../../@core/services';
-import { LayoutService } from '../../../@core/utils';
+import { LayoutService } from '../../../@core/services/layout.service';
 import { TranslationBaseComponent } from '../../../@shared/language-base';
-import { ChangeDetectorRef } from '@angular/core';
 import { OrganizationTeamStore } from '../../../@core/services/organization-team-store.service';
 import { QuickActionsComponent } from '../../../@shared/dialogs/quick-actions/quick-actions.component';
 
@@ -61,7 +48,6 @@ import { QuickActionsComponent } from '../../../@shared/dialogs/quick-actions/qu
 	templateUrl: './header.component.html'
 })
 export class HeaderComponent extends TranslationBaseComponent implements OnInit, OnDestroy, AfterViewInit {
-
 	isEmployee = false;
 	isElectron: boolean = environment.IS_ELECTRON;
 	isDemo: boolean = environment.DEMO;
@@ -247,12 +233,12 @@ export class HeaderComponent extends TranslationBaseComponent implements OnInit,
 			(event: KeyboardEvent, handler: HotkeysEvent) => {
 				const pressedKeys: string = this.pressedKeysToString(handler);
 				// -- prevent triggering shortcuts when typing in NbOptions (searchable inputs)
-				if ((event.target as Element).localName === "nb-option") return;
+				if ((event.target as Element).localName === 'nb-option') return;
 				if (this.shortcutsMap.has(pressedKeys)) {
 					// -- close dialog if open when using shortcuts and prevent closing dialog by pressing unregistered keys
 					this.quickActionsRef &&
-						pressedKeys !== this.defaultShortcuts.quickActions.toLowerCase() &&
-						this.shortcutsMap.has(pressedKeys)
+					pressedKeys !== this.defaultShortcuts.quickActions.toLowerCase() &&
+					this.shortcutsMap.has(pressedKeys)
 						? this.quickActionsRef.close()
 						: null;
 					if (handler.scope === 'defaultShortcuts' && this.shortcutsMap.has(pressedKeys)) {
@@ -275,7 +261,7 @@ export class HeaderComponent extends TranslationBaseComponent implements OnInit,
 
 	private registerShortcut(keys: string, action: () => void) {
 		this.shortcutsMap.set(keys.toLowerCase(), action);
-		hotkeys(keys, 'defaultShortcuts', () => { });
+		hotkeys(keys, 'defaultShortcuts', () => {});
 		hotkeys.setScope('defaultShortcuts');
 	}
 
@@ -623,521 +609,521 @@ export class HeaderComponent extends TranslationBaseComponent implements OnInit,
 			// Divider (Accounting)
 			...(this.store.hasAnyPermission(PermissionsEnum.INVOICES_EDIT, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_INVOICE'),
-						icon: 'file-text-outline',
-						link: 'pages/accounting/invoices/add',
-						data: {
-							action: 'createInvoice'
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.createInvoice),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_INVOICE'),
+							icon: 'file-text-outline',
+							link: 'pages/accounting/invoices/add',
+							data: {
+								action: 'createInvoice'
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.createInvoice),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(!!this.user.employee
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.RECEIVED_INVOICES'),
-						icon: {
-							icon: 'file-invoice-dollar',
-							pack: 'font-awesome'
-						},
-						link: 'pages/accounting/invoices/received-invoices',
-						data: {
-							action: 'receivedInvoices'
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.receivedInvoices),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.RECEIVED_INVOICES'),
+							icon: {
+								icon: 'file-invoice-dollar',
+								pack: 'font-awesome'
+							},
+							link: 'pages/accounting/invoices/received-invoices',
+							data: {
+								action: 'receivedInvoices'
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.receivedInvoices),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_INCOMES_EDIT, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_INCOME'),
-						icon: 'plus-circle-outline',
-						link: 'pages/accounting/income',
-						data: {
-							action: 'createIncome'
-						},
-						queryParams: {
-							openAddDialog: true
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.createIncome),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_INCOME'),
+							icon: 'plus-circle-outline',
+							link: 'pages/accounting/income',
+							data: {
+								action: 'createIncome'
+							},
+							queryParams: {
+								openAddDialog: true
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.createIncome),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_EXPENSES_EDIT, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_EXPENSE'),
-						icon: 'minus-circle-outline',
-						link: 'pages/accounting/expenses',
-						data: {
-							action: 'createExpense'
-						},
-						queryParams: {
-							openAddDialog: true
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.createExpense),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_EXPENSE'),
+							icon: 'minus-circle-outline',
+							link: 'pages/accounting/expenses',
+							data: {
+								action: 'createExpense'
+							},
+							queryParams: {
+								openAddDialog: true
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.createExpense),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 
 			...(this.store.hasAnyPermission(PermissionsEnum.ESTIMATES_EDIT, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_ESTIMATE'),
-						icon: 'file-outline',
-						link: 'pages/accounting/invoices/estimates/add',
-						data: {
-							action: 'createEstimate'
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.createEstimate),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_ESTIMATE'),
+							icon: 'file-outline',
+							link: 'pages/accounting/invoices/estimates/add',
+							data: {
+								action: 'createEstimate'
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.createEstimate),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(!!this.user.employee
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.RECEIVED_ESTIMATES'),
-						icon: {
-							icon: 'file-invoice',
-							pack: 'font-awesome'
-						},
-						link: 'pages/accounting/invoices/received-estimates',
-						data: {
-							action: 'receivedEstimates'
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.receivedEstimates),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.RECEIVED_ESTIMATES'),
+							icon: {
+								icon: 'file-invoice',
+								pack: 'font-awesome'
+							},
+							link: 'pages/accounting/invoices/received-estimates',
+							data: {
+								action: 'receivedEstimates'
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.receivedEstimates),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_PAYMENT_ADD_EDIT, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_PAYMENT'),
-						icon: 'credit-card-outline',
-						link: 'pages/accounting/payments',
-						data: {
-							action: 'createPayment'
-						},
-						queryParams: {
-							openAddDialog: true
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.createPayment),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_PAYMENT'),
+							icon: 'credit-card-outline',
+							link: 'pages/accounting/payments',
+							data: {
+								action: 'createPayment'
+							},
+							queryParams: {
+								openAddDialog: true
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.createPayment),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			// Divider (Organization)
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_EMPLOYEES_EDIT, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.ADD_EMPLOYEE'),
-						icon: 'people-outline',
-						link: 'pages/employees',
-						data: {
-							action: 'addEmployee'
-						},
-						queryParams: {
-							openAddDialog: true
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.addEmployee),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.ADD_EMPLOYEE'),
+							icon: 'people-outline',
+							link: 'pages/employees',
+							data: {
+								action: 'addEmployee'
+							},
+							queryParams: {
+								openAddDialog: true
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.addEmployee),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_INVENTORY_PRODUCT_EDIT, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.ADD_INVENTORY'),
-						icon: 'inbox-outline',
-						link: 'pages/organization/inventory/create',
-						data: {
-							action: 'addInventory'
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.addInventory),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.ADD_INVENTORY'),
+							icon: 'inbox-outline',
+							link: 'pages/organization/inventory/create',
+							data: {
+								action: 'addInventory'
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.addInventory),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_EQUIPMENT_EDIT, PermissionsEnum.ALL_ORG_EDIT) ||
-				!!this.user.employee
+			!!this.user.employee
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.ADD_EQUIPMENT'),
-						icon: 'pantone-outline',
-						link: 'pages/organization/equipment',
-						data: {
-							action: 'addEquipment'
-						},
-						queryParams: {
-							openAddDialog: true
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.addEquipment),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.ADD_EQUIPMENT'),
+							icon: 'pantone-outline',
+							link: 'pages/organization/equipment',
+							data: {
+								action: 'addEquipment'
+							},
+							queryParams: {
+								openAddDialog: true
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.addEquipment),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.ADD_VENDOR'),
-						icon: 'car-outline',
-						link: 'pages/organization/vendors',
-						data: {
-							action: 'addVendor'
-						},
-						queryParams: {
-							openAddDialog: true
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.addVendor),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.ADD_VENDOR'),
+							icon: 'car-outline',
+							link: 'pages/organization/vendors',
+							data: {
+								action: 'addVendor'
+							},
+							queryParams: {
+								openAddDialog: true
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.addVendor),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.ADD_DEPARTMENT'),
-						icon: 'briefcase-outline',
-						link: 'pages/organization/departments',
-						data: {
-							action: 'addDepartment'
-						},
-						queryParams: {
-							openAddDialog: true
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.addDepartment),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.ADD_DEPARTMENT'),
+							icon: 'briefcase-outline',
+							link: 'pages/organization/departments',
+							data: {
+								action: 'addDepartment'
+							},
+							queryParams: {
+								openAddDialog: true
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.addDepartment),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 
 			// Divider (Project Management)
 			...(this.store.hasAnyPermission(PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_TEAM'),
-						icon: 'people-outline',
-						link: 'pages/organization/teams',
-						data: {
-							action: 'createTeam'
-						},
-						queryParams: {
-							openAddDialog: true
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.createTeam),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_TEAM'),
+							icon: 'people-outline',
+							link: 'pages/organization/teams',
+							data: {
+								action: 'createTeam'
+							},
+							queryParams: {
+								openAddDialog: true
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.createTeam),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasPermission(PermissionsEnum.ORG_TASK_ADD)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_TASK'),
-						icon: 'archive-outline',
-						link: 'pages/tasks/dashboard',
-						data: {
-							action: 'createTask'
-						},
-						queryParams: {
-							openAddDialog: true
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.createTask),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_TASK'),
+							icon: 'archive-outline',
+							link: 'pages/tasks/dashboard',
+							data: {
+								action: 'createTask'
+							},
+							queryParams: {
+								openAddDialog: true
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.createTask),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_PROJECT_ADD, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_PROJECT'),
-						icon: 'color-palette-outline',
-						link: 'pages/organization/projects/create',
-						data: {
-							action: 'createProject'
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.createProject),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_PROJECT'),
+							icon: 'color-palette-outline',
+							link: 'pages/organization/projects/create',
+							data: {
+								action: 'createProject'
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.createProject),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_TASK_VIEW, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.VIEW_TASKS'),
-						icon: 'calendar-outline',
-						link: 'pages/tasks/dashboard',
-						data: {
-							action: 'viewTasks'
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.viewTasks),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.VIEW_TASKS'),
+							icon: 'calendar-outline',
+							link: 'pages/tasks/dashboard',
+							data: {
+								action: 'viewTasks'
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.viewTasks),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_TASK_VIEW, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.VIEW_TEAM_TASKS'),
-						icon: 'layers-outline',
-						link: 'pages/tasks/team',
-						data: {
-							action: 'viewTeamTasks'
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.viewTeamTasks),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.VIEW_TEAM_TASKS'),
+							icon: 'layers-outline',
+							link: 'pages/tasks/team',
+							data: {
+								action: 'viewTeamTasks'
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.viewTeamTasks),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 
 			// Divider (Jobs)
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_CANDIDATES_EDIT, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_CANDIDATE'),
-						icon: 'person-add-outline',
-						link: 'pages/employees/candidates',
-						data: {
-							action: 'createCandidate'
-						},
-						queryParams: {
-							openAddDialog: true
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.createCandidate),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_CANDIDATE'),
+							icon: 'person-add-outline',
+							link: 'pages/employees/candidates',
+							data: {
+								action: 'createCandidate'
+							},
+							queryParams: {
+								openAddDialog: true
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.createCandidate),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_PROPOSALS_EDIT, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_PROPOSAL'),
-						icon: 'paper-plane-outline',
-						link: 'pages/sales/proposals/register',
-						data: {
-							action: 'createProposal'
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.createProposal),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_PROPOSAL'),
+							icon: 'paper-plane-outline',
+							link: 'pages/sales/proposals/register',
+							data: {
+								action: 'createProposal'
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.createProposal),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_CONTRACT_EDIT, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_CONTRACT'),
-						icon: 'file-text-outline',
-						link: 'pages/integrations/upwork',
-						data: {
-							action: 'createContract'
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.createContract),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_CONTRACT'),
+							icon: 'file-text-outline',
+							link: 'pages/integrations/upwork',
+							data: {
+								action: 'createContract'
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.createContract),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			// Divider (Contacts)
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_CONTACT_EDIT, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_LEAD'),
-						icon: 'shield-outline',
-						link: 'pages/contacts/leads',
-						data: {
-							action: 'createLead'
-						},
-						queryParams: {
-							openAddDialog: true
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.createLead),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_LEAD'),
+							icon: 'shield-outline',
+							link: 'pages/contacts/leads',
+							data: {
+								action: 'createLead'
+							},
+							queryParams: {
+								openAddDialog: true
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.createLead),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_CONTACT_EDIT, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_CUSTOMER'),
-						icon: 'person-done-outline',
-						link: 'pages/contacts/customers',
-						data: {
-							action: 'createCustomer'
-						},
-						queryParams: {
-							openAddDialog: true
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.createCustomer),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_CUSTOMER'),
+							icon: 'person-done-outline',
+							link: 'pages/contacts/customers',
+							data: {
+								action: 'createCustomer'
+							},
+							queryParams: {
+								openAddDialog: true
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.createCustomer),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.ORG_CONTACT_EDIT, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_CLIENT'),
-						icon: 'person-outline',
-						link: 'pages/contacts/clients',
-						data: {
-							action: 'createClient'
-						},
-						queryParams: {
-							openAddDialog: true
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.createClient),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.CREATE_CLIENT'),
+							icon: 'person-outline',
+							link: 'pages/contacts/clients',
+							data: {
+								action: 'createClient'
+							},
+							queryParams: {
+								openAddDialog: true
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.createClient),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(!!this.user.employee
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.VIEW_CLIENTS'),
-						icon: 'person-outline',
-						link: 'pages/contacts/clients',
-						data: {
-							action: 'viewClients'
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.viewClients),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.VIEW_CLIENTS'),
+							icon: 'person-outline',
+							link: 'pages/contacts/clients',
+							data: {
+								action: 'viewClients'
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.viewClients),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			// Divider (Time Tracking)
 			...(this.store.hasAnyPermission(PermissionsEnum.TIMESHEET_EDIT_TIME, PermissionsEnum.ALL_ORG_EDIT)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.TIME_LOG'),
-						icon: 'clock-outline',
-						link: 'pages/employees/timesheets/daily',
-						data: {
-							action: 'timeLog'
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.timeLog),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.TIME_LOG'),
+							icon: 'clock-outline',
+							link: 'pages/employees/timesheets/daily',
+							data: {
+								action: 'timeLog'
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.timeLog),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.TIME_TRACKER)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.VIEW_APPOINTMENTS'),
-						icon: 'calendar-outline',
-						link: 'pages/employees/appointments',
-						data: {
-							action: 'viewAppointments'
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.viewAppointments),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.VIEW_APPOINTMENTS'),
+							icon: 'calendar-outline',
+							link: 'pages/employees/appointments',
+							data: {
+								action: 'viewAppointments'
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.viewAppointments),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(this.store.hasAnyPermission(PermissionsEnum.TIME_TRACKER)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.VIEW_TIME_ACTIVITY'),
-						icon: 'activity-outline',
-						link: 'pages/employees/activity/time-activities',
-						data: {
-							action: 'viewTimeActivity'
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.viewTimeActivity),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.VIEW_TIME_ACTIVITY'),
+							icon: 'activity-outline',
+							link: 'pages/employees/activity/time-activities',
+							data: {
+								action: 'viewTimeActivity'
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.viewTimeActivity),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(!!this.user.employee
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.START_TIMER'),
-						icon: 'play-circle-outline',
-						hidden: !this.isEmployee || this.isElectron,
-						data: {
-							action: this.actions.START_TIMER // -- Start the timer
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.startTimer),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.START_TIMER'),
+							icon: 'play-circle-outline',
+							hidden: !this.isEmployee || this.isElectron,
+							data: {
+								action: this.actions.START_TIMER // -- Start the timer
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.startTimer),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: []),
 			...(!this.store.hasAnyPermission(PermissionsEnum.TIMESHEET_EDIT_TIME)
 				? [
-					{
-						title: this.getTranslation('QUICK_ACTIONS_MENU.STOP_TIMER'),
-						icon: 'pause-circle-outline',
-						hidden: !this.isEmployee || this.isElectron,
-						data: {
-							action: this.actions.STOP_TIMER // -- Stop the timer
-						},
-						badge: {
-							text: this.formatShortcut(this.defaultShortcuts.stopTimer),
-							status: 'control'
+						{
+							title: this.getTranslation('QUICK_ACTIONS_MENU.STOP_TIMER'),
+							icon: 'pause-circle-outline',
+							hidden: !this.isEmployee || this.isElectron,
+							data: {
+								action: this.actions.STOP_TIMER // -- Stop the timer
+							},
+							badge: {
+								text: this.formatShortcut(this.defaultShortcuts.stopTimer),
+								status: 'control'
+							}
 						}
-					}
-				]
+				  ]
 				: [])
 		];
 	}
@@ -1192,5 +1178,5 @@ export class HeaderComponent extends TranslationBaseComponent implements OnInit,
 		this.isCollapse = event;
 	}
 
-	ngOnDestroy() { }
+	ngOnDestroy() {}
 }
