+++ apps/gauzy/src/app/@theme/theme.module.ts
@@ -21,12 +21,7 @@ import {
 import { NbEvaIconsModule } from '@nebular/eva-icons';
 import { NbSecurityModule } from '@nebular/security';
 
-import {
-	FooterComponent,
-	HeaderComponent,
-	LayoutSelectorComponent,
-	SearchInputComponent
-} from './components';
+import { FooterComponent, HeaderComponent, LayoutSelectorComponent, SearchInputComponent } from './components';
 import { Pipes } from './pipes';
 import {
 	OneColumnLayoutComponent,
@@ -50,17 +45,13 @@ import { WindowModeBlockScrollService } from './services';
 import { LanguagesService, OrganizationsService, UsersOrganizationsService, UsersService } from '../@core/services';
 import { HeaderSelectorsModule } from './components/header/selectors/selectors.module';
 import { EmployeeSelectorsModule } from './components/header/selectors/employee/employee.module';
-import { SelectorService } from '../@core/utils/selector.service';
+import { SelectorService } from '../@core/services/selector.service';
 import { TimeTrackerModule } from '../@shared/time-tracker/time-tracker.module';
 import { ProjectSelectModule } from '../@shared/project-select/project-select.module';
 import { TranslateModule } from '../@shared/translate/translate.module';
 import { ThemeSidebarModule } from './components/theme-sidebar/theme-sidebar.module';
 import { MatRippleModule } from '@angular/material/core';
-import {
-	NbAccordionModule,
-	NbToggleModule,
-	NbCardModule
-} from '@nebular/theme';
+import { NbAccordionModule, NbToggleModule, NbCardModule } from '@nebular/theme';
 import { GauzyLogoComponent } from './components/gauzy-logo/gauzy-logo.component';
 import { UserMenuComponent } from './components/user-menu/user-menu.component';
 import { UserComponent } from './components/user/user.component';
@@ -70,7 +61,7 @@ import { DateRangePickerModule } from './components/header/selectors/date-range-
 import { WorkspacesModule } from '../@shared/workspaces/workspaces.module';
 import { PopupComponent } from './components/popup/popup.component';
 import { SidebarMenuModule } from '../@shared/sidebar-menu/sidebar-menu.module';
-import { DirectivesModule } from "../@shared/directives/directives.module";
+import { DirectivesModule } from '../@shared/directives/directives.module';
 import { TimeTrackerStatusModule } from '../@shared/time-tracker/components/time-tracker-status/time-tracker-status.module';
 import { ThemeLanguageSelectorService } from './components/theme-sidebar/theme-settings/components/theme-language-selector.service';
 import { TeamSelectModule } from '../@shared/team-select/team-select.module';
